@@ -167,7 +168,7 @@
                     <version>${maven-javadoc-plugin.version}</version>
                     <configuration>
                         <quiet>true</quiet>
-                        <source>1.8</source>
+                        <source>21</source>
                     </configuration>
                 </plugin>
                 <plugin>
